Move NewRequisitionType to common type file
