--- client/packages/requisitions/src/RequestRequisition/ListView/types.ts
@@ -1,4 +1,0 @@
-export enum NewRequisitionType {
-  General,
-  Program,
-}
