+++ client/packages/requisitions/src/types.ts
@@ -1,0 +1,4 @@
+export enum NewRequisitionType {
+  General,
+  Program,
+}
